Adding Extent Report functionality to test suite.

+++ src/test/java/resources/stepDefinitions/MasterHooks.java
@@ -1,8 +1,11 @@
 package resources.stepDefinitions;
 
 
+import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
 import resources.pageObjects.BasePage;
 
 
@@ -16,9 +19,28 @@ public class MasterHooks extends BasePage {
 
     @After("@End")
 
-    public void teardown() {
-        driver.manage().deleteAllCookies();
+    public void tearDownAndScreenshotOnFailure(Scenario scenario) {
+        try {
+            if (driver != null && scenario.isFailed()) {
+                scenario.embed(((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES), "image/png");
+                BasePage.captureScreenshot();
+                driver.manage().deleteAllCookies();
+                driver.quit();
+                driver = null;
+            }
+            if (driver != null) {
+                driver.manage().deleteAllCookies();
+                driver.quit();
+                driver = null;
+            }
+        } catch (Exception e) {
+            System.out.println("Methods failed: tearDownAndScreenshotOnFailure, Exception: " + e.getMessage());
+        }
+
+
+
+/*        driver.manage().deleteAllCookies();
         driver.close();
-        driver = null;
+        driver = null;*/
     }
 }
